This is a very basic implementation of the top items, the view still needs a lot of work

+++ QuickMenu/Menu.h
@@ -25,6 +25,6 @@
 
 -(int)topItemCount;
 
--(id)itemForTopPosition:(int)position;
+-(id)itemForTopPosition:(long)position;
 
 @end
